requirements.txt
Flask==2.0.2
Pillow==8.3.2
pytesseract==0.3.8
joblib==1.1.0
Werkzeug==2.2.2